Moving unit tests around
